Nokia plug-in URL and tiler API updates
Updated URLs for places and map tiler to current ones. Added URL
management for additional hosts for hybrid, terrain and satellite maps.
Added new languages and map types available in the new tiler API.

Bug fixes for code evaluating copyrights text to show. Now only
copyrights related to viewed maps are shown and each name is shown only
once.

Updated drawing of copyright information to show HERE logo and
copyrights text side by side. This is to avoid having logo jumping
up and down.

+++ src/plugins/geoservices/nokia/qgeotiledmappingmanagerengine_nokia.h
@@ -78,7 +78,7 @@ public:
 
     virtual QGeoMapData *createMapData();
     QString evaluateCopyrightsText(const QGeoMapType::MapStyle mapStyle,
-                                   const int zoomLevel,
+                                   const qreal zoomLevel,
                                    const QSet<QGeoTileSpec> &tiles);
 
 public Q_SLOTS:
